fin_cander/requirements.txt
asgiref==3.7.2
black==23.7.0
bootstrap4==0.1.0
boto3
click==8.1.6
colorama==0.4.6
crispy-bootstrap4==2022.1
Django==3.2.0
django-crispy-forms==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
django-storages
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
pip==23.1.2
platformdirs==3.10.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
setuptools==65.5.1
sqlparse==0.2.4
tzdata==2023.3
wheel==0.38.4
django-cors-headers
djongo
mongoengine
pylint-mongoengine
pymongo==3.12.3
django-utils-six